Обновление

+++ code/game/gamemodes/cult/runes.dm
@@ -120,8 +120,8 @@
 			animate(src)
 
 /obj/effect/rune/examine(mob/user)
-	. = ..()
-	if(iscultist(user))
+	..()
+	if(iscultist(user) || isghost(user))
 		to_chat(user, "This is \a [cultname] rune.")
 
 /obj/effect/rune/attackby(obj/item/I, mob/living/user)
@@ -158,7 +158,7 @@
 
 /obj/effect/rune/proc/get_cultists()
 	. = list()
-	for(var/mob/living/M in range(1))
+	for(var/mob/living/M in range(1, src))
 		if(iscultist(M))
 			. += M
 
@@ -431,12 +431,20 @@
 	cultname = "defile"
 
 /obj/effect/rune/defile/cast(mob/living/user)
+	if(!is_station_turf(get_turf(src)))
+		to_chat(user, SPAN_DANGER("This place is too powerless for any defiling!"))
+		return
 	speak_incantation(user, "Ia! Ia! Zasan therium viortia!")
 	for(var/turf/T in range(1, src))
 		if(T.holy)
-			T.holy = 0
+			T.holy = FALSE
 		else
 			T.cultify()
+		
+	var/area/A = get_area(src)
+	if(A && !isspace(A))
+		A.holy = FALSE
+
 	visible_message("<span class='warning'>\The [src] embeds into the floor and walls around it, changing them!</span>", "You hear liquid flow.")
 	qdel(src)
 
@@ -681,6 +689,9 @@
 
 /obj/effect/rune/massdefile/cast(mob/living/user)
 	var/list/mob/living/cultists = get_cultists()
+	if(!is_station_turf(get_turf(src)))
+		to_chat(user, SPAN_DANGER("This place is too powerless for any defiling!"))
+		return
 	if(cultists.len < 3)
 		to_chat(user, "<span class='warning'>You need three cultists around this rune to make it work.</span>")
 		return fizzle(user)
@@ -689,9 +700,14 @@
 			M.say("Ia! Ia! Zasan therium viortia! Razan gilamrua kioha!")
 		for(var/turf/T in range(5, src))
 			if(T.holy)
-				T.holy = 0
+				T.holy = FALSE
 			else
 				T.cultify()
+
+	var/area/A = get_area(src)
+	if(A && !isspace(A))
+		A.holy = FALSE
+
 	visible_message("<span class='warning'>\The [src] embeds into the floor and walls around it, changing them!</span>", "You hear liquid flow.")
 	qdel(src)
 
@@ -798,15 +814,18 @@
 /obj/effect/rune/blood_boil
 	cultname = "blood boil"
 	strokes = 4
+	var/is_used = FALSE
 
 /obj/effect/rune/blood_boil/cast(mob/living/user)
 	var/list/mob/living/cultists = get_cultists()
 	if(cultists.len < 3)
 		return fizzle()
+	if(is_used)
+		return fizzle()
 
 	for(var/mob/living/M in cultists)
 		M.say("Dedo ol[pick("'","`")]btoh!")
-
+	is_used = TRUE
 	var/list/mob/living/previous = list()
 	var/list/mob/living/current = list()
 	while(cultists.len >= 3)
@@ -827,6 +846,7 @@
 		previous = current.Copy()
 		current.Cut()
 		sleep(10)
+	is_used = FALSE
 
 /* Tier NarNar runes */
 
@@ -839,6 +859,9 @@
 
 /obj/effect/rune/tearreality/cast(mob/living/user)
 	if(!GLOB.cult.allow_narsie)
+		return
+	if(!is_station_turf(get_turf(src)))
+		to_chat(user, SPAN("cult", "Our Deity does not have enough influence on this place to be summoned there!"))
 		return
 	if(the_end_comes)
 		to_chat(user, "<span class='cult'>You are already summoning! Be patient!</span>")
@@ -913,7 +936,7 @@
 	var/obj/item/weapon/paper/target
 	var/tainted = 0
 	for(var/obj/item/weapon/paper/P in get_turf(src))
-		if(!P.info)
+		if(P.is_clean())
 			target = P
 			break
 		else
